add precon subpackage for nystroem preconditioner
